requirements.txt
glfw==1.4.0
numpy==1.13.3
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.0
